@@ -8,19 +8,19 @@
         "d3": "^7.1.1",
         "object-hash": "^3.0.0",
         "openseadragon": "^4.0.0",
-        "react": "^17.0.1",
-        "react-dom": "^17.0.1",
-        "react-loading-icons": "^1.0.6",
-        "react-modal": "^3.12.1",
-        "react-responsive": "^9.0.0-beta.10",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "^5.0.0",
-        "react-select": "^5.4.0",
-        "react-tabs": "^3.2.2",
-        "react-toastify": "^8.0.3",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-loading-icons": "^1.1.0",
+        "react-modal": "^3.16.1",
+        "react-responsive": "^9.0.2",
+        "react-router-dom": "^5.3.4",
+        "react-scripts": "^5.0.1",
+        "react-select": "^5.7.2",
+        "react-tabs": "^6.0.1",
+        "react-toastify": "^9.1.2",
         "reactjs-popup": "^2.0.5",
-        "recoil": "^0.5.0",
-        "recoil-nexus": "^0.3.8",
+        "recoil": "^0.7.7",
+        "recoil-nexus": "^0.4.0",
         "validator": "^13.6.0"
     },
     "devDependencies": {
@@ -29,11 +29,10 @@
         "@types/node": "^16.10.3",
         "@types/object-hash": "^2.2.1",
         "@types/openseadragon": "^3.0.5",
-        "@types/react": "^17.0.27",
-        "@types/react-dom": "^17.0.9",
+        "@types/react": "^18.2.0",
+        "@types/react-dom": "^18.2.1",
         "@types/react-responsive": "^8.0.5",
-        "@types/react-router-dom": "^5.3.1",
-        "@types/react-tabs": "^2.3.3",
+        "@types/react-router-dom": "^5.3.3",
         "@types/validator": "^13.6.3",
         "autoprefixer": "^10.4.4",
         "eslint-config-prettier": "^8.3.0",
