@@ -1,16 +1,18 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import { RecoilRoot } from "recoil";
 import RecoilNexus from "recoil-nexus";
 import App from "./App";
 import "./index.css";
+import { createRoot } from "react-dom/client";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container!);
+
+root.render(
     <React.StrictMode>
         <RecoilRoot>
             <RecoilNexus />
             <App />
         </RecoilRoot>
-    </React.StrictMode>,
-    document.getElementById("root")
+    </React.StrictMode>
 );
